Euclid's number to find the GCD
Euclid used an algorithm to find the greatest common denominator between two values in which the smaller value was divided into the larger value until the remainder was less than the smaller value. The smaller value was divided by the remainder until the number left was less than the remainder. This number is the GCD.